fix: Change classname on lecture info modal
